+++ HAL/Src/AnalogIn.cpp
@@ -252,7 +252,7 @@ void AnalogIn::detachSamplingProgressCallback()
  * @brief Initializes ADC peripheral and creates a FreeRTOS task for handling incoming DMA ADC reads
  * @note Must be called in main() in order for ADC to work.
 */
-void AnalogIn::initialize(uint16_t sample_rate)
+void AnalogIn::initialize(uint32_t sample_rate)
 {
     // initialize DMA
     __HAL_RCC_DMA2_CLK_ENABLE(); /* DMA controller clock enable */
